--- a/conftest.py
+++ b/conftest.py
@@ -5,6 +5,7 @@ def pytest_addoption(parser):
     parser.addoption('--fixme', action='store_true')
     parser.addoption('--gui', action='store_true')
     parser.addoption('--show', action='store_true')
+    parser.addoption('--slow', action='store_true')
     parser.addoption('--tomcat', action='store_true')
     parser.addoption('--web-tests', action='store_true')
     parser.addoption('--weird', action='store_true')