--- a/setup.cfg
+++ b/setup.cfg
@@ -43,7 +43,7 @@ single-quotes = true
 
 [tool:pytest]
 minversion = 5.4
-addopts = -p no:doctest --xdoctest --xdoctest-style=google -r fEs
+addopts = -v -p no:doctest --xdoctest --xdoctest-style=google -r fEs
 testpaths =
     wbia
 filterwarnings =