--- a/wbia/algo/verif/deploy.py
+++ b/wbia/algo/verif/deploy.py
@@ -447,7 +447,7 @@ class Deployer(object):
                 * PUBLISH TO /media/hdd/PUBLIC/models/pairclf
 
         Example:
-            >>> # xdoctest: +REQUIRES(module:wbia_cnn)
+            >>> # xdoctest: +REQUIRES(module:wbia_cnn, --slow)
             >>> from wbia.algo.verif.vsone import *  # NOQA
             >>> params = dict(sample_method='random')
             >>> pblm = OneVsOneProblem.from_empty('PZ_MTEST', **params)