--- a/wbia/web/job_engine.py
+++ b/wbia/web/job_engine.py
@@ -352,6 +352,7 @@ def get_job_metadata(ibs, jobid):
         python -m wbia.web.job_engine --exec-get_job_metadata:0 --fg
 
     Example:
+        >>> # xdoctest: +REQUIRES(--slow)
         >>> # WEB_DOCTEST
         >>> from wbia.web.job_engine import *  # NOQA
         >>> import wbia