I ran `pytest --durations=10` to find out the slowest 10 tests:

```
599.58s call     wbia/algo/graph/mixin_matching.py::InfrLearning.learn_evaluation_verifiers:0
235.24s call     wbia/algo/hots/tests/test_lnbnn.py::test_lnbnn
112.55s call     wbia/algo/smk/smk_pipeline.py::SMK.match_single:0
96.87s call     wbia/algo/graph/mixin_matching.py::CandidateSearch.find_lnbnn_candidate_edges:0
57.68s call     wbia/algo/hots/chip_match.py::_ChipMatchVisualization.show_single_namematch:0
35.15s call     wbia/expt/test_result.py::TestResult.get_truth2_prop:0
25.27s call     wbia/algo/smk/inverted_index.py::InvertedAnnots.compute_word_weights:0
18.80s call     wbia/expt/test_result.py::TestResult.get_cfgx_groupxs:0
10.69s call     wbia/algo/graph/mixin_loops.py::InfrLoops.main_gen:0
10.25s call     wbia/algo/graph/mixin_matching.py::AnnotInfrMatching._cm_training_pairs:0
```

I added `# xdoctest: +REQUIRES(--slow)` to the tests so our tests
(without `--slow`) run faster.

After doing this a few times, I marked 26 tests as slow.  (Every time I
marked a test as slow, a different test becomes slow)